Extracted inner classes from WorldPanel
